Run authorize on Docker
