+++ authentication/app/client.py
@@ -2,7 +2,7 @@ import pika
 import threading
 import json
 
-connection = pika.BlockingConnection(pika.ConnectionParameters(host='localhost'))
+connection = pika.BlockingConnection(pika.ConnectionParameters(host='192.168.99.100'))
 channel = connection.channel()
 
 
@@ -14,16 +14,13 @@ channel.queue_declare(queue='auth-service')
 channel.queue_declare(queue='response-queue')
 channel.basic_consume(queue='response-queue', on_message_callback=callback, auto_ack=True)
 
-
 thread = threading.Thread(
         target=channel.start_consuming,
         daemon=True)
 thread.start()
 
 while True:
-
     body = input("Insert Packet: ")
-    print(body)
     packet = json.loads(body)
     packet['response_queue_name'] = 'response-queue'
     body = json.dumps(packet).encode('utf-8')
