+++ authentication/requirements.txt
@@ -1,5 +1,5 @@
 pika
 requests
-Flask
+flask
 oauthlib
-pyOpenSSL
+pyopenssl
